+++ src/routes/RequestsRoute.js
@@ -36,6 +36,7 @@ import {
   DefaultMCLRowFormatter,
   NoValue,
   MCLPagingTypes,
+  Icon,
 } from '@folio/stripes/components';
 import {
   deparseFilters,
@@ -1548,7 +1549,9 @@ class RequestsRoute extends React.Component {
               to={`${this.props.location.pathname}?layer=create`}
               onClick={onToggle}
             >
-              <FormattedMessage id="stripes-smart-components.new" />
+              <Icon icon="plus-sign">
+                <FormattedMessage id="stripes-smart-components.new" />
+              </Icon>
             </Button>
           </IfPermission>
           {csvReportPending ?
@@ -1565,7 +1568,9 @@ class RequestsRoute extends React.Component {
                 this.exportData();
               }}
             >
-              <FormattedMessage id="ui-requests.exportSearchResultsToCsv" />
+              <Icon icon="download">
+                <FormattedMessage id="ui-requests.exportSearchResultsToCsv" />
+              </Icon>
             </Button>
           }
           {
@@ -1584,10 +1589,12 @@ class RequestsRoute extends React.Component {
                     this.exportExpiredHoldsToCSV();
                   }}
                 >
-                  <FormattedMessage
-                    id="ui-requests.exportExpiredHoldShelfToCsv"
-                    values={{ currentServicePoint: servicePointName }}
-                  />
+                  <Icon icon="download">
+                    <FormattedMessage
+                      id="ui-requests.exportExpiredHoldShelfToCsv"
+                      values={{ currentServicePoint: servicePointName }}
+                    />
+                  </Icon>
                 </Button>
                 <PrintButton
                   buttonStyle="dropdownItem"
@@ -1604,10 +1611,12 @@ class RequestsRoute extends React.Component {
                   }}
                   onAfterPrint={this.onAfterPrintForPrintButton}
                 >
-                  <FormattedMessage
-                    id="ui-requests.printPickSlips"
-                    values={{ sp: servicePointName }}
-                  />
+                  <Icon icon="print">
+                    <FormattedMessage
+                      id="ui-requests.printPickSlips"
+                      values={{ sp: servicePointName }}
+                    />
+                  </Icon>
                 </PrintButton>
                 <PrintButton
                   buttonStyle="dropdownItem"
@@ -1636,10 +1645,12 @@ class RequestsRoute extends React.Component {
                   }
                   onAfterPrint={this.onAfterPrintForPrintButton}
                 >
-                  <FormattedMessage
-                    id="ui-requests.printPickSlipsSelected"
-                    values={{ sp: servicePointName }}
-                  />
+                  <Icon icon="print">
+                    <FormattedMessage
+                      id="ui-requests.printPickSlipsSelected"
+                      values={{ sp: servicePointName }}
+                    />
+                  </Icon>
                 </PrintButton>
               </>
           }
@@ -1659,10 +1670,12 @@ class RequestsRoute extends React.Component {
                     contentRef={this.searchSlipsPrintContentRef}
                     onBeforeGetContent={() => this.onBeforeGetContentForPrintButton(onToggle)}
                   >
-                    <FormattedMessage
-                      id="ui-requests.printSearchSlips"
-                      values={{ sp: servicePointName }}
-                    />
+                    <Icon icon="print">
+                      <FormattedMessage
+                        id="ui-requests.printSearchSlips"
+                        values={{ sp: servicePointName }}
+                      />
+                    </Icon>
                   </PrintButton>
               }
             </>