+++ src/routes/RequestsRoute.js
@@ -235,7 +235,6 @@ export const urls = {
 export const getListFormatter = (
   {
     getRowURL,
-    setURL,
   },
   {
     intl,
@@ -288,7 +287,7 @@ export const getListFormatter = (
     ? <FormattedMessage id={requestStatusesTranslations[rq.status]} />
     : <NoValue />),
   'type': rq => <FormattedMessage id={requestTypesTranslations[rq.requestType]} />,
-  'title': rq => <TextLink to={getRowURL(rq.id)} onClick={() => setURL(rq.id)}>{(rq.instance ? rq.instance.title : DEFAULT_FORMATTER_VALUE)}</TextLink>,
+  'title': rq => <TextLink to={getRowURL(rq.id)}>{(rq.instance ? rq.instance.title : DEFAULT_FORMATTER_VALUE)}</TextLink>,
   'year': rq => getFormattedYears(rq.instance?.publication, DEFAULT_DISPLAYED_YEARS_AMOUNT),
   'callNumber': rq => effectiveCallNumber(rq.item),
   'servicePoint': rq => get(rq, 'pickupServicePoint.name', DEFAULT_FORMATTER_VALUE),
@@ -660,7 +659,6 @@ class RequestsRoute extends React.Component {
       errorModalData: {},
       servicePointId: '',
       requests: [],
-      selectedId: '',
       selectedRows: {},
       titleLevelRequestsFeatureEnabled,
       createTitleLevelRequestsByDefault,
@@ -1075,20 +1073,6 @@ class RequestsRoute extends React.Component {
     } = this.props;
 
     return `${path}/view/${id}${search}`;
-  }
-
-  setURL = (id) => {
-    this.setState({
-      selectedId: id,
-    });
-  }
-
-  resultIsSelected = ({ item }) => item.id === this.state.selectedId;
-
-  viewRecordOnCollapse = () => {
-    this.setState({
-      selectedId: null,
-    });
   }
 
   getHelperResourcePath = (helper, id) => `circulation/requests/${id}`;
@@ -1518,7 +1502,6 @@ class RequestsRoute extends React.Component {
     const resultsFormatter = getListFormatter(
       {
         getRowURL: this.getRowURL,
-        setURL: this.setURL,
       },
       {
         intl,
@@ -1741,7 +1724,7 @@ class RequestsRoute extends React.Component {
                 copies: { max: 95 },
               }}
               columnMapping={columnLabels}
-              resultsRowClickHandlers={false}
+              hasRowClickHandlers={false}
               resultsFormatter={resultsFormatter}
               resultRowFormatter={DefaultMCLRowFormatter}
               newRecordInitialValues={initialValues}
@@ -1772,11 +1755,9 @@ class RequestsRoute extends React.Component {
                 isEcsTlrSettingReceived,
                 isEcsTlrSettingEnabled,
               }}
-              viewRecordOnCollapse={this.viewRecordOnCollapse}
               viewRecordPerms="ui-requests.view"
               newRecordPerms="ui-requests.create"
               renderFilters={this.renderFilters}
-              resultIsSelected={this.resultIsSelected}
               onFilterChange={this.handleFilterChange}
               sortableColumns={['requestDate', 'title', 'year', 'itemBarcode', 'callNumber', 'type', 'requestStatus',
                 'position', 'servicePoint', 'requester', 'requesterBarcode', 'retrievalServicePoint', ...(isProxyAvailable ? ['proxy'] : []), 'copies', 'printed']}