+++ src/routes/RequestsRoute.test.js
@@ -269,8 +269,6 @@ SearchAndSort.mockImplementation(jest.fn(({
   onFilterChange,
   parentResources,
   renderFilters,
-  resultIsSelected,
-  viewRecordOnCollapse,
   customPaneSub,
   columnMapping,
   resultsFormatter,
@@ -280,12 +278,6 @@ SearchAndSort.mockImplementation(jest.fn(({
     onDuplicate(parentResources.records.records[0]);
     buildRecordsForHoldsShelfReport();
     massageNewRecord({});
-    resultIsSelected({
-      item: {
-        id: 'id',
-      },
-    });
-    viewRecordOnCollapse();
   };
   return (
     // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
@@ -1277,7 +1269,6 @@ describe('RequestsRoute', () => {
 
   describe('getListFormatter', () => {
     const getRowURLMock = jest.fn(id => id);
-    const setURLMock = jest.fn(id => id);
     const getPrintContentRefMock = jest.fn(id => id);
     const isPrintableMock = jest.fn(id => id);
     const toggleRowSelectionMock = jest.fn(id => id);
@@ -1286,7 +1277,6 @@ describe('RequestsRoute', () => {
     const listFormatter = getListFormatter(
       {
         getRowURL: getRowURLMock,
-        setURL: setURLMock,
       },
       {
         intl,
@@ -1497,7 +1487,6 @@ describe('RequestsRoute', () => {
       it('should render "TextLink" with correct props', () => {
         const expectedProps = {
           to: requestWithData.id,
-          onClick: expect.any(Function),
         };
 
         render(listFormatter.title(requestWithData));