+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add "Retrieval service point" filter. Refs UIREQ-1190.
 * Include `Retrieval service point` column in csv exports. Refs UIREQ-1191.
 * Add icons to requests action menu. Refs UIREQ-1116.
+* Revise "Enter" button on New request page. Refs UIREQ-1114.
 
 ## [11.0.2] (https://github.com/folio-org/ui-requests/tree/v11.0.2) (2024-12-10)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v11.0.1...v11.0.2)