+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add optional column "Retrieval service point" to requests search list. Refs UIREQ-1188.
 * Cleanup retrieval service point implementation from deprecated folder. Refs UIREQ-1211.
 * Add "Retrieval service point" filter. Refs UIREQ-1190.
+* Include `Retrieval service point` column in csv exports. Refs UIREQ-1191.
 
 ## [11.0.2] (https://github.com/folio-org/ui-requests/tree/v11.0.2) (2024-12-10)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v11.0.1...v11.0.2)