+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-requests
 
+## [11.1.0] IN PROGRESS
+
+* Add `requests ID` as a search parameter for search box. Refs UIREQ-1232.
+
 ## [11.0.3] (https://github.com/folio-org/ui-requests/tree/v11.0.3) (2025-01-10)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v11.0.2...v11.0.3)